+++ AddOns/Cell/Locales/zhCN.lua
@@ -487,6 +487,10 @@ L["Built-in Spells"] = "内置法术"
 L["Highlight Type"] = "高亮类型"
 L["Shape"] = "形状"
 L["To show shield value, |cffff2727Glyph of Power Word: Shield|r is required"] = "需要有|cffff2727真言术：盾雕文|r才能显示盾值"
+L["Cast By"] = "来源"
+L["Me"] = "我"
+L["Anyone"] = "任何人"
+L["Others"] = "其他人"
 
 L["Click to preview"] = "点击预览"
 L["Debug Mode"] = "调试模式"
@@ -751,6 +755,24 @@ L["Click to view older changelogs"] = "点击查看远古更新记录"
 -- <br/>
 
 L["CHANGELOGS"] = [[
+    <h1>r201-release (Nov 9, 2023, 05:04 GMT+8)</h1>
+    <p>* 更新副本减益：阿梅达希尔，梦境之愿。</p>
+    <p>* 更新指示器：血量文字。</p>
+    <br/>
+
+    <h1>r200-release (Nov 4, 2023, 08:04 GMT+8)</h1>
+    <p>* 更新部分团队工具的配置选项。</p>
+    <p>* 更新“减伤（自身）”。</p>
+    <p>* 更新繁中。</p>
+    <br/>
+        
+    <h1>r199-release (Oct 21, 2023, 15:40 GMT+8)</h1>
+    <p>* 更新“减伤”指示器的内置法术。</p>
+    <p>* 为增益类型的自定义指示器添加了“来源”选项。</p>
+    <p>* 修复团队工具。</p>
+    <p>+ 代码片段变量：CELL_SHOW_RAID_PET_OWNER_NAME。</p>
+    <br/>
+
     <h1>r198-release (Oct 7, 2023, 06:54 GMT+8)</h1>
     <p>* 更新指示器：目标计数，护盾条。</p>
     <br/>