+++ AddOns/Cell/CHANGELOG.md
@@ -1,4 +1,4 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r197-release...8eaa4d855902bf89773d05c9224ce1214d696f2e)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r200-release...35b0ccfd96956e313858d81e4d3da795e963fa34)
 
-- Update Target Counter
-- Update position options of Shield Bar indicator
+- Update Amirdrassil debuffs
+- Update Health Text